+++ modules/pybind11/src/bindings/deserialize/deserialize.h
@@ -16,6 +16,7 @@
 #include "ods_occupancy_grid_v1.h"
 #include "o3d_paramters.h"
 #include "ods_polar_occupancy_grid_v1.h"
+#include "ods_extrinsic_calibration_correction_v1.h"
 
 void
 bind_deserialize_struct(pybind11::module_& m)
@@ -27,6 +28,7 @@ bind_deserialize_struct(pybind11::module_& m)
   bind_struct_odsoccupancygridv1(m);
   bind_struct_o3d_paramters(m);
   bind_struct_odspolaroccupancygridv1(m);
+  bind_struct_odsextrinsiccalibrationcorrectionv1(m);
 }
 
 #endif // IFM3D_PYBIND_BINDING_DESERIALIZE